[feat/#21]Refactor: Separate navbar
