+++ src/components/Submit.tsx
@@ -1,46 +1,12 @@
-import React, { FC, useState, useCallback } from "react";
-import { Link, useNavigate } from "react-router-dom";
-import axios from "axios";
+import React, { FC } from "react";
 import SubmitterList from "../props/SubmitterList";
 import "../style/home.css";
+import Navbar from "./Navbar";
 const Submit: FC = () => {
-  // nav bar login btn
-  const navigate = useNavigate();
-  const logout = () => {
-    navigate("/", {
-      state: {},
-    });
-  };
-  const mypage = () => {
-    navigate("/mypage", {
-      state: {},
-    });
-  };
-
   return (
     <>
       <div className="background">
-        <div
-          className="navbar"
-          style={{
-            display: "flex",
-            justifyContent: "space-between",
-            paddingLeft: 50,
-            paddingRight: 50,
-          }}
-        >
-          <div>
-            <h3 className="m-3">Logo</h3>
-          </div>
-          <div>
-            <button className="navbtn" onClick={logout}>
-              로그아웃
-            </button>
-            <button className="navbtn" onClick={mypage}>
-              마이페이지
-            </button>
-          </div>
-        </div>
+        <Navbar />
         <div className="card mb-6 mt-3 profile-submitter-box">
           <div
             style={{
