+++ src/style/Navbar.css
@@ -1,0 +1,7 @@
+.navbar {
+    height: 87px;
+    left: 0px;
+    top: 0px;
+    width: 100%;
+    background: #383839;
+  }
